+++ Questionable/Data/QuestData.cs
@@ -191,6 +191,8 @@ internal sealed class QuestData
         _quests = quests.ToDictionary(x => x.QuestId, x => x);
 
         AddPreviousQuest(new(2898), new(2922));
+        AddPreviousQuest(new(5008), new(4933));
+        AddPreviousQuest(new(5012), new(4959));
         // workaround because the game doesn't require completion of the CT questline through normal means
         AddPreviousQuest(new(425), new(495));