+++ Questionable/Windows/QuestWindow.cs
@@ -137,7 +137,7 @@ internal sealed class QuestWindow : LWindow, IPersistableWindowConfig
             {
                 Icon = FontAwesomeIcon.Heart,
                 IconOffset = TitleBarIconOffset,
-                Click = _ => Process.Start(new ProcessStartInfo { FileName = "https://github.com/sponsors/alydevs", UseShellExecute = true }),
+                Click = _ => Process.Start(new ProcessStartInfo { FileName = "https://ko-fi.com/alydev", UseShellExecute = true }),
                 Priority = TitleBarButtonPriority,
                 ShowTooltip = () =>
                 {